+++ examples/Commands/src/App.ts
@@ -169,6 +169,12 @@ namespace LiteMolPluginInstance {
     addButton('Components: Without Log', () => {
         plugin.instance.setComponents(NoLogComponents);
     });
+    addButton('Fog Factor = 0.5', () => {
+        plugin.context.scene.scene.updateOptions({ fogFactor: 0.5 });
+    });
+    addButton('Fog Factor = 1.0', () => {
+        plugin.context.scene.scene.updateOptions({ fogFactor: 1.0 });
+    });
     
     addSeparator();
 
@@ -218,7 +224,7 @@ namespace LiteMolPluginInstance {
         let ambQ =  Query.residues({ name: 'REA' }).ambientResidues(5); // adjust the radius
 
         let id = '1cbs:REA'
-        let url = `https://webchem.ncbr.muni.cz/CoordinateServer/1cbs/ligandInteraction?name=REA`; // here you will fill in the full server etc ...
+        let url = `https://cs.litemol.org/1cbs/ligandInteraction?name=REA`; // here you will fill in the full server etc ...
         let action = Transform.build()
             .add(plugin.context.tree.root, Transformer.Data.Download, { url, type: 'String', id })
             .then(Transformer.Data.ParseCif, { id }, { isBinding: true })
@@ -473,7 +479,7 @@ namespace LiteMolPluginInstance {
                 'molecule.model.defaultQuery': `residues({ name: 'ALA' })`,
                 'molecule.model.defaultAssemblyName': '1', 
                 'molecule.coordinateStreaming.defaultId': '1jj2',
-                'molecule.coordinateStreaming.defaultServer': 'https://webchem.ncbr.muni.cz/CoordinateServer/',
+                'molecule.coordinateStreaming.defaultServer': 'https://cs.litemol.org/',
                 'molecule.coordinateStreaming.defaultRadius': 10,
                 'density.defaultVisualBehaviourRadius': 5
             },