+++ src/lib/Plugin/Views/Visualization/Viewport.tsx
@@ -27,6 +27,7 @@ namespace LiteMol.Plugin.Views.Visualization {
             if (this.state.showSceneOptions) {             
                 options = <div className='lm-viewport-controls-scene-options lm-control'>                    
                     <Controls.Toggle onChange={v => this.controller.setState({ enableFog: v })} value={state.enableFog!} label='Fog' />
+                    <Controls.Slider label='Fog Factor' min={0.1} max={1} step={0.01} onChange={v => this.controller.setState({ fogFactor: v }) } value={state.fogFactor!} />
                     <Controls.Slider label='FOV' min={30} max={90} onChange={v => this.controller.setState({ cameraFOV: v }) } value={state.cameraFOV!} />
                     <Controls.Slider label='Camera Speed' min={1} max={10} step={0.01} onChange={v => this.controller.setState({ cameraSpeed: v }) } value={state.cameraSpeed!} />
                     <Controls.ToggleColorPicker color={state.clearColor!} label='Background' position='below' onChange={c => this.controller.setState({ clearColor: c }) } />