+++ src/lib/Visualization/CHANGELOG.md
@@ -1,5 +1,6 @@
 ﻿# 1.7.9
 * Fixed a shader bug that broke Chrome 75 compat.
+* Support Camera Fog Factor.
 
 # 1.7.8
 * Relaxed condition for switching to "backbone only" mode.